requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dateutil==2.9.0.post0

python-dotenv
pandas
numpy

streamlit
toml
requests